+++ src/app/core/core.module.ts
@@ -1,7 +1,8 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { HeaderComponent } from './components/header/header.component';
 import { RouterModule } from '@angular/router';
+
+import { HeaderComponent } from './components/header/header.component';
 
 @NgModule({
     declarations: [
@@ -13,6 +14,6 @@ import { RouterModule } from '@angular/router';
     ],
     exports: [
         HeaderComponent,
-    ]
+    ],
 })
 export class CoreModule { }