+++ changelog.md
@@ -1,7 +1,19 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.0.3] - 2022-08-08
+## [0.3.0] - 2022-08-08
+### Changed
+- the way to track an empty cart.
+- increase/decrease implementation.
+
+### Added
+- ability to clear all products in cart.
+- id generator.
+- local storage service.
+- zoom product image in cart.
+- log additional info from optional services on product page.
+
+## [0.2.0] - 2022-08-08
 ### Changed
 - total cost moved to cart service.
 - detection strategy.
@@ -14,14 +26,14 @@ All notable changes to this project will be documented in this file.
 - the ability to increase ang decrease products amount.
 - highlight item in cart on hover.
 
-## [0.0.2] - 2022-08-07
+## [0.1.1] - 2022-08-07
 ### Changed
 - unsibscribe.
 - remove null initialization.
 - refactored getProductById function.
 - remove lazy-load for root component module.
 
-## [0.0.1] - 2022-07-10
+## [0.1.0] - 2022-07-10
 ### Added
 - header component for navigation (according task 8).
 - product component to display full information about product (according task 3 and 4). 
@@ -29,6 +41,7 @@ All notable changes to this project will be documented in this file.
 - product list item component to display list of products (according task 5).
 - cart list compnent to display items in cart and their total price (according task 7, 8).
 
-[0.0.1]: https://github.com/illiaZakharov/Angular-34/tree/hw1
-[0.0.2]: https://github.com/illiaZakharov/Angular-34/tree/fix/task1-comments
-[0.0.3]: https://github.com/illiaZakharov/Angular-34/tree/hw2
+[0.1.0]: https://github.com/illiaZakharov/Angular-34/tree/hw1
+[0.1.1]: https://github.com/illiaZakharov/Angular-34/tree/fix/task1-comments
+[0.2.0]: https://github.com/illiaZakharov/Angular-34/tree/hw2
+[0.3.0]: https://github.com/illiaZakharov/Angular-34/tree/hw3